+++ src/config.ts
@@ -1,0 +1,7 @@
+import * as vscode from "vscode";
+
+export const CONFIG_SECTION = "claudeTokens";
+
+export function config(): vscode.WorkspaceConfiguration {
+  return vscode.workspace.getConfiguration(CONFIG_SECTION);
+}